refactor(desktop): extract browser panel from main process
